working (no dispatching yet), working on a memory allocation system

+++ core.cpp
@@ -43,7 +43,7 @@ int Board::run() {
     int exitCode = 0;
     //init clock & SDL2 timings
     static constexpr int LOOP_SPEED_HZ = 200000;           // estimated from SDL2 bottleneck, calibrated for 36 MHz
-    static constexpr int TARGET_CLOCK_SPEED_HZ = 36000000; // target clock speed overall
+    static constexpr long long TARGET_CLOCK_SPEED_HZ = 36000000; // overall target clock speed
     static constexpr int STEPS_PER_LOOP = TARGET_CLOCK_SPEED_HZ / LOOP_SPEED_HZ;
     SDL_Event e;
     clock.resetCycles(); //set clock cycles to zero @ the start of a new process
